requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-openai
langchain-ollama
